Small changes, unit tests last step
